stock_ticker/requirements.txt
pytest
requests
vcrpy